chore(types): move inline types from API services to central types module
- Move CommunicationsParams to types/communication.ts
- Move PublicGroupDto, PublicGroupSearchParams to types/group.ts
- Move BatchFilterParams, BatchListResponse to types/giving.ts
- Move UploadFileOptions to types/files.ts
- Update API services to import from @/types
- Re-export types for backward compatibility

Closes #358
