+++ src/web/src/types/giving.ts
@@ -161,3 +161,38 @@ export interface PersonLookupDto {
   fullName: string;
   email?: string;
 }
+
+// ============================================================================
+// Batch Filter and Response Types
+// ============================================================================
+
+/**
+ * Filter parameters for batch search
+ */
+export interface BatchFilterParams {
+  /** Filter by status (Open, Closed, Posted) */
+  status?: string;
+  /** Filter by campus IdKey */
+  campusIdKey?: string;
+  /** Filter by start date (inclusive) */
+  startDate?: string;
+  /** Filter by end date (inclusive) */
+  endDate?: string;
+  /** Page number (1-based) */
+  page?: number;
+  /** Items per page */
+  pageSize?: number;
+}
+
+/**
+ * Response envelope for paginated batch results
+ */
+export interface BatchListResponse {
+  data: ContributionBatchDto[];
+  meta: {
+    page: number;
+    pageSize: number;
+    totalCount: number;
+    totalPages: number;
+  };
+}
