+++ tools/graph/backend-graph.json
@@ -1,6 +1,6 @@
 {
   "version": "1.0",
-  "generated_at": "2026-01-02T20:38:19.642112+00:00",
+  "generated_at": "2026-01-02T20:57:03.107725+00:00",
   "entities": {
     "Attendance": {
       "name": "Attendance",
